// src-31052018-currently-working/app/core/enums/page-context-types.ts
/**
 * An enumeration for the contextual pages within GPR, currently used with {@link SortOptionsService} and for
 * storing user preferences
 */
export enum PageContextTypes {
  USER_HOME = 'GPR Home Page',
  CUSTOMER_HOME = 'Customer Home Page',
  NAV_MENU_CUSTOMERS = 'Navigation Menu - Customers',
  NAV_MENU_PLANS = 'Navigation Menu - Plans',
  HISTORY = 'HISTORY'
}